requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
ultralytics
python-multipart
torch
torchvision
numpy
scikit-learn
pillow
python-dotenv
